pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "loop-generator"
version = "0.1.0"
description = "AI-powered loop generator with Python GenAI integration and C++ DSP engine"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "soundfile>=0.12.1",
    "google-genai>=1.0.0",
]

[project.scripts]
loop-generator = "cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["python*", "cli*", "config*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
